+++ CHANGELOG.md
@@ -6,11 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Next v2.0.0-dev.8 2023-09-06 (Draft)
+## Next v2.0.0-dev.9 2023-09-08 (Draft)
 
 Work in progress, see top of README.md
 
 Commit
+
+Added more debug tools
+
+- In raw-message panel, added checkbox to view both raw IRC server message as text and parsed message object as JSON.
+- In raw-message panel, added feature to submit a sample IRC server message to the IRC message parser while observing browser dev tools console.
+
+Commit d170d18
 
 - Fixed button on server list form call wrong function (copy/paste error)